Simplified pagination controls.
Instead of having most of the pagination logic in the template, moved
it to the controller instead: introduced the mojology_page() function,
callable from the template.

Now the template only has to figure out whether to display the
prev/next buttons, the rest is taken care of by the application.