ajaxification, part one
Added a new URL handler, "/log/<logid>/dyn", which returns only the
table with dynamic values. This is, of course, in its own template,
which is used by log.html aswell.

Furthermore, Javascript code was added to intercept clicks on the
table rows, and route them to a function that will load any data
necessary, and do the appropriate magic to show it.