Added "logs" to the main menu.
Explicitly added "logs" to the main menu, now that we have a couple of
entries, and clicking on the logo wasn't obvious enough anymore.

As a side effect, clicking on the logo brings one to the main
dashboard (which, at the moment, is just an internal redirect to the
browser dashboard).