db: Redesigned how the database layout config works.
Implemented mojology.db.DefaultLayout() which presents an interface
which can describe the database layout, aswell as control some aspects
of what gets displayed and how: It can filter out any of the keys, it
can describe what the main headers are.

It can also apply some minor modifications to the data, if so need be.