mojology: Moved the connection code to a decorator
Having the connection code in a decorator, instead of doing it before
each request, we can have routes that do not connect to mongodb: such
as the error handlers, which will be displayed when mongodb is not
accessible.

This way, we can finally display HTTP 503 when the database is not
available.